Borysov/quiz-boot/src/main/java/com/quiz/borysov/quizboot/service/UserService.java
package com.quiz.borysov.quizboot.service;

import com.quiz.borysov.quizboot.dao.UserDao;
import com.quiz.borysov.quizboot.entity.User;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.stereotype.Component;
import org.springframework.stereotype.Service;

import java.security.MessageDigest;
import java.security.NoSuchAlgorithmException;
import java.sql.SQLException;

@Service
public class UserService {
    private static final Logger logger = LoggerFactory.getLogger(UserService.class);

    @Autowired
    private UserDao userDao;

    public User getUserByLoginAndPass(String login, String password) {

        return (User) userDao.findAll();
    }

    public void registerUser(User newUser) {
        newUser.setPassword(new BCryptPasswordEncoder().encode(newUser.getPassword()));
        userDao.save(newUser);
    }


}
